refactor: update import paths and rename CreateProductParams to AddProduct
